Trilinos Master Merge PR Generator: Auto PR created to promote from master_merge_20181017_012007 branch to master (#3650)
* MiniEM: Move problem definitions to input xml

* MiniEM: Enable MPI_4 test of MueLu+Epetra

* Panzer: Add integrator 'BasisTimesTensorTimesVector'

* MiniEM: Solver config updates

* MiniEM: Add random current

* Panzer:  Update Integrator_GradBasisCro... (#1892)

Create a compile-time argument checking constructor that supports
CONTRIBUTES-style evaluation.  Create a ParameterList constructor that
delegates to the main one for the sake of backwards compatibility.  Also
Kokkos-ify the evaluateFields() routine.

* MiniEM: make conductivity a tensor

* MueLu RefMaxwell: Read "verbosity" from ParameterList

* MiniEM: Better checks of command line params, prepare for Scalar=complex

* Framework: Add debugging message to PR test driver

Adding an echo to print out the contents of JOB_BASE_NAME for debugging.

* Framework: Guard SEMS_MPI_NAME in PR driver

The PR Driver script prints out SEMS_MPI_NAME using
`${SEMS_MPI_NAME:?}`, which will throw an error if
`SEMS_MPI_NAME` is nonexistent or empty... which it is
for the SERIAL build.

* Framework: Syncing comment to what Jim had

* Update PullRequestLinuxDriver.sh

Generalizing SERIAL conditional to any serial build, instead of the specific one we are adding as suggested by Mark H.

* revert disable of panzer Curl test on waterman #3340

* MueLu: fix bug in Phase3 aggregation

The number of unaggregated nodes was being decremented in the wrong
spot.  This should fix github issue #3556 and some of the MueLu test
failures under CUDA reported in github issue #2674.

* Fix view example with Intel compiler for issue #3639

This test is failing for Intel builds for some reason.  It only fails
with Intel, and only with high optimization.  Trivial changes to the
kernel make it pass, like adding a print statement or the temporary I
added here.  Smells like an optimization bug.