Have Kokkos TriBITS build set compiler options as target properties (#11545)
With the TriBITS modernization refactoring (TriBITSPub/TriBITS#299) and the
generalizated handling of intenral and external packages
(TriBITSPub/TriBITS#63), we need packages like Kokkos to set critical compiler
options as target properties so that they will be exported in the generated
IMPORTED targets of the Kokkos<Subpkg>Targets.cmake file.

This is needed, for example, to pass some critical compiler flags from the
pre-installed Kokkos to downstream CMake configures of KokkosKernels and the
rest of Trilinos (see #11545).